Refactor server into modular architecture
